Organize project structure with dedicated directories
