+++ setup.py
@@ -3,15 +3,7 @@ from setuptools import setup, find_packages
 setup(
     name="qualitylab",
     version="0.1.0",
-    py_modules=[
-        "build_time",
-        "build_quantity",
-        "defects",
-        "feature_engineering",
-        "cli",
-        "spreadsheets",
-        "streamlit_app",
-    ],
+    packages=find_packages(),
     install_requires=[
         "pandas",
         "scikit-learn",
@@ -24,7 +16,7 @@ setup(
     ],
     entry_points={
         "console_scripts": [
-            "qualitylab=cli:cli",
+            "qualitylab=qualitylab.cli:cli",
         ]
     }
 )
